requirements.txt
torch
torchvision
pillow
pre-commit
pytest
pytest-cov
